Changed all film_id columns to larger integer type Changed film.length to unsigned Changed film.release_year to small integer because mysql year type does not support <1901

+++ database/migrations/2020_01_22_213808_create_inventory_table.php
@@ -15,7 +15,7 @@ class CreateInventoryTable extends Migration
     {
         Schema::create('inventory', function (Blueprint $table) {
             $table->mediumIncrements('inventory_id');
-            $table->unsignedSmallInteger('film_id');
+            $table->unsignedInteger('film_id');
             $table->unsignedTinyInteger('store_id');
             $table->timestamps();
