+++ database/migrations/2020_01_22_215109_create_film_category_table.php
@@ -14,7 +14,7 @@ class CreateFilmCategoryTable extends Migration
     public function up()
     {
         Schema::create('film_category', function (Blueprint $table) {
-            $table->unsignedSmallInteger('film_id');
+            $table->unsignedInteger('film_id');
             $table->unsignedTinyInteger('category_id');
             $table->timestamps();
